% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{config.help}
\alias{config.help}
\title{Function to access external helps about configurations format or other related information}
\usage{
config.help(name = NULL, show_all_names = FALSE)
}
\arguments{
\item{name}{Name or number of helps}

\item{show_all_names}{Show all urls name}
}
\description{
Function to access external helps about configurations format or other related information
}
\examples{
 config.help()
\dontrun{
 config.help(1)
 config.help('ini_git_search')
}
}
