% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{config.list.merge}
\alias{config.list.merge}
\title{Merge list file (From config package), list.right will overwrite the element also existed in list.left}
\usage{
config.list.merge(list.left = list(), list.right = list())
}
\arguments{
\item{list.left}{One list be merged left}

\item{list.right}{One list be merged right}
}
\value{
A list
}
\description{
Merge list file (From config package), list.right will overwrite the element also existed in list.left
}
\examples{
config.json <- system.file('extdata', 'config.json', package='configr')
list.left <- list()
list.right <- eval.config(file = config.json)
config.list.merge(list.left, list.right)
list.left <- list(a=c(123,456))
list.right <- list(a=c(4,5,6)) 
config.list.merge(list.left, list.right)
}
\seealso{
\code{\link[config]{merge}} call in this function
}
