% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del.R
\name{config.sections.del}
\alias{config.sections.del}
\title{Delete sections in config, just do config[sections] <- NULL}
\usage{
config.sections.del(config, sections)
}
\arguments{
\item{config}{a list of config (eg. generated by read.config)}

\item{sections}{Sections that need to be deleted}
}
\value{
A list of config
}
\description{
Delete sections in config, just do config[sections] <- NULL
}
\examples{
config.json <- system.file('extdata', 'config.json', package = 'configr')
config <- read.config(config.json, file.type = 'json')
config <- config.sections.del(config, 'default')
}
