% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configR-package.R
\docType{package}
\name{configR-package}
\alias{configR}
\alias{configR-package}
\title{configR: R Conviguration Package}
\description{
Setup and configure an optimal R development environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/configR}
  \item \url{https://jimbrig.github.io/configR/}
  \item Report bugs at \url{https://github.com/jimbrig/configR/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
