% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configr.R
\docType{package}
\name{configr}
\alias{configr}
\alias{configr-package}
\title{configr package implements the YAML parser, 
JSON parser and INI parser for R setting and writing of configuration file.}
\description{
configr package implements the YAML parser, 
JSON parser and INI parser for R setting and writing of configuration file.
}

