% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{eval.config.merge.groups}
\alias{eval.config.merge.groups}
\title{Merge config file groups}
\usage{
eval.config.merge.groups(file = "config.json", groups = NULL,
  encoding = getOption("encoding"))
}
\arguments{
\item{file, }{Configuration file to read from (defaults to 'config.json') 
'JSON/INI/YAML format file.}

\item{groups, }{Set need be merged parameter sets, eg. groups=c('default', 'version')}

\item{encoding, }{Encoding of filepath parameter, will default to system encoding if not specifield}
}
\value{
a list
}
\description{
Merge config file groups
}
\examples{
config.json <- system.file('extdata', 'config.json', package='configr')
eval.config.merge.groups(config.json)
}

