% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{is.config.active}
\alias{is.config.active}
\title{Test active configuration}
\usage{
is.config.active(config)
}
\arguments{
\item{config}{Configuration name}
}
\value{
Logical indicating whether the specified configuration is active
}
\description{
Check whether a configuration group is currently active
}
\examples{
is.config.active('default')
}
