% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{is.configfile.active}
\alias{is.configfile.active}
\title{Test active configuration file}
\usage{
is.configfile.active(config.file)
}
\arguments{
\item{config.file}{Configuration filename}
}
\value{
Logical indicating whether the specified configuration file is active
}
\description{
Check whether a configuration file is currently active
}
\examples{
is.configfile.active('config.cfg')
}
