% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configtype.R
\name{is.ini.file}
\alias{is.ini.file}
\title{Function to check wheather file is ini format}
\usage{
is.ini.file(file, encoding = getOption("encoding"))
}
\arguments{
\item{file}{filename to be coerced or tested}

\item{encoding}{Encoding of filepath parameter, will default to system encoding if not specifield}
}
\value{
TRUE/FALSE
}
\description{
Function to check wheather file is ini format
}
\examples{
config.json <- system.file('extdata', 'config.json', package='configr')
config.ini <- system.file('extdata', 'config.ini', package='configr')
config.yaml <- system.file('extdata', 'config.yaml', package='configr')
print(is.ini.file(config.ini))
print(is.ini.file(config.json))
print(is.ini.file(config.yaml))
}
\seealso{
\code{\link{is.json.file}}/\code{\link{is.yaml.file}}
}

