% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configtype.R
\name{is.yaml.file}
\alias{is.yaml.file}
\title{Function to check wheather file is YAML format}
\usage{
is.yaml.file(file, encoding = getOption("encoding"))
}
\arguments{
\item{file, }{filename to be coerced or tested}

\item{encoding, }{Encoding of filepath parameter, will default to system encoding if not specifield}
}
\value{
TRUE/FALSE
}
\description{
Function to check wheather file is YAML format
}
\examples{
config.json <- system.file('extdata', 'config.json', package='configr')
config.ini <- system.file('extdata', 'config.ini', package='configr')
config.yaml <- system.file('extdata', 'config.yaml', package='configr')
print(is.yaml.file(config.yaml))
print(is.json.file(config.json))
print(is.ini.file(config.ini))
}

