% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_config_dir.R
\name{r_config_dir}
\alias{r_config_dir}
\title{r_config_dir}
\usage{
r_config_dir(...)
}
\arguments{
\item{...}{Path elements passed as character strings to append to \code{R_CONFIG_DIR}}
}
\value{
normalized path to the \code{R_CONFIG_DIR}
}
\description{
Helper function for retrieving and settings the user's custom \code{R_CONFIG_DIR}.
}
\details{
This function appends path elements to your default \code{R_CONFIG_DIR}. To find
\code{R_CONFIG_DIR} it will first attempt to search for the \code{R_CONFIG_DIR} environment
variable, and if unset defaults to \verb{~/.config/R} on all operating systems
(i.e. \verb{\%USERPROFILE\%\\.config\\R} on Windows).
}
\examples{
# retrieve path to .Rprofile (i.e. ~/.config/R/.Rprofile)
r_config_dir(".Rprofile")

# retrieve path to your "profile" configuration directory
# (i.e. ~/.config/R/profile)
r_config_dir("profile")
}
