% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.config}
\alias{read.config}
\title{Read from the file (JSON/INI/YAML be supported), retreiving all values as a list.}
\usage{
read.config(file = "config.json", encoding = getOption("encoding"))
}
\arguments{
\item{file, }{Configuration file to read from (defaults to 'config.json') 
'JSON/INI/YAML format file.}

\item{encoding, }{Encoding of filepath parameter, will default to system encoding if not specifield}
}
\value{
All values as a list.
}
\description{
Read from the file (JSON/INI/YAML be supported), retreiving all values as a list.
}
\examples{
config.json <- system.file('extdata', 'config.json', package='configr')
config <- read.config(file=config.json)
}

