% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write.config}
\alias{write.config}
\title{Write a config list in a file (JSON/YAML/INI)}
\usage{
write.config(config.dat, file.path, write.type = "ini",
  encoding = getOption("encoding"))
}
\arguments{
\item{config.dat, }{a list of config (eg. generated by read.config)}

\item{file.path, }{file to write}

\item{write.type, }{json/ini/yaml}

\item{encoding, }{Encoding of filepath parameter, will default to system
encoding if not specifield}
}
\value{
a list
}
\description{
Write a config list in a file (JSON/YAML/INI)
}
\examples{
list.test <- list(a=c(123,456))
write.config(list.test, '/tmp/test.yaml', write.type = 'yaml')
}

