% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configtype.R
\name{get.config.type}
\alias{get.config.type}
\title{Get config file type:json/ini/yaml retreiving json/ini/yaml}
\usage{
get.config.type(file = "config.json", encoding = getOption("encoding"))
}
\arguments{
\item{file, }{Configuration file to read from (defaults to 'config.json') 
'JSON/INI/YAML format file.}

\item{encoding, }{Encoding of filepath parameter, will default to system encoding if not specifield}
}
\value{
json/ini/yaml
}
\description{
Get config file type:json/ini/yaml retreiving json/ini/yaml
}
\examples{
config.json <- system.file('extdata', 'config.json', package='configr')
get.config.type(file=config.json)
}

